+++ pkg/util/async/runner_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package async
 
 import (
 	"fmt"
