+++ pkg/util/rlimit/rlimit_linux.go
@@ -1,7 +1,7 @@
 // +build linux
 
 /*
-Copyright 2016 The Kubernetes Authors All rights reserved.
+Copyright 2016 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
